checkpoint: v2 realtime GUI stack + deep UI audit hardening
Checkpoints the uncommitted v2-only migration (vendored Applio realtime
engine, RealtimeSession control layer, PySide6/QML GUI, offline model-merge,
niepian presets) and lands this session's UI audit fixes on top.

UI audit (5-dimension adversarial review, 17 confirmed defects fixed, all
GUI/control-layer, all faithful-carrier-contract-safe):

- backend threads: shutdown waits unbounded on an in-flight load/merge so the
  process can no longer abort on exit (QThread destroyed while still running);
  threads deleteLater + refs cleared (no per-load leak); busy-guards on
  startOrStop/reloadModel stop combo/f0 re-entry from silently dropping loads.
- backend correctness: sidReset signal re-syncs the sid spinbox on every load
  (fixes the same-speaker-count swap lying); load failure surfaced once (ERROR
  state only); merged-model profile path is frozen-safe (models/<name>.pth).
- pure modules: model_default_params derives formant_on (a saved gender shift
  is no longer silently disabled/overwritten); drop the dead formant_qfrency
  key; save_model_profile clamps + rejects NaN/inf; blend_weights self-guards
  equal key-sets (torch import moved below validation).
- Main.qml: device combos survive a refresh (selName/onModelChanged restore,
  no silent fall-back to the wrong device); f0 default pinned; outName dedup;
  formant checkbox initialised from the profile.
- efficiency: metrics poll only while RUNNING; drop the dead presetsChanged
  signal (presets is constant); single shared build_audio_config builder.

.gitignore: also exclude model weights (.bin/.onnx/.safetensors) so the 360MB
contentvec embedder is never committed.

Verify: pytest 150 passed; check_qml.py 0 warnings + LOAD OK; smoke_ui_backend
PASS; throwaway torch-path blend smoke PASS.

+++ src/safety/guard.py
@@ -38,7 +38,7 @@ def dbfs_peak(audio: np.ndarray) -> float:
     peak = float(np.max(np.abs(a)))
     if peak <= 0.0 or not np.isfinite(peak):
         return DBFS_SILENCE_FLOOR
-    return 20.0 * np.log10(peak)
+    return float(20.0 * np.log10(peak))   # native float: np.log10 returns np.float64
 
 
 def dbfs_rms(audio: np.ndarray) -> float:
@@ -50,7 +50,7 @@ def dbfs_rms(audio: np.ndarray) -> float:
     if mean_sq <= 0.0 or not np.isfinite(mean_sq):
         return DBFS_SILENCE_FLOOR
     rms = np.sqrt(mean_sq)
-    return 20.0 * np.log10(rms)
+    return float(20.0 * np.log10(rms))    # native float: np.log10 returns np.float64
 
 
 @dataclass(frozen=True)
